container condense files
