+++ src/Components/containerBuild.js
@@ -1,12 +1,43 @@
 import React from "react";
-import { Group, Rect } from "react-konva";
+import { Group, Rect, Circle } from "react-konva";
 import ContainerColor from "./containerColor";
-import { ContainerTitle } from "./containerTitle";
-import { ContainerContent } from "./containerContent";
-import { ContainerFooter } from "./containerFooter";
+import { Html } from "react-konva-utils";
 
 function ContainerBuild() {
   const { colorFill, clickColor } = ContainerColor();
+
+  const TitleInput = {
+    width: 160,
+    height: 25,
+    margin: 10,
+    border: "none",
+    borderBottom: "2px solid white",
+    padding: "10px",
+    background: "none",
+    resize: "none",
+    color: "white",
+    fontSize: "20px",
+    fontFamily: "Helvetica",
+    fontWeight: "bold",
+    // outline: "1px solid red",
+  };
+
+  const ContentInput = {
+    width: 160,
+    height: 140,
+    position: "absolute",
+    top: 55,
+    margin: 10,
+    border: "none",
+    borderBottom: "2px solid white",
+    padding: "10px",
+    background: "none",
+    resize: "none",
+    color: "white",
+    fontSize: "16px",
+    fontFamily: "Helvetica",
+    // outline: "1px solid red",
+  };
 
   return (
     <Group>
@@ -24,12 +55,38 @@ function ContainerBuild() {
       />
 
       {/* //* container title text area */}
-      <ContainerTitle />
+      <Html>
+        <textarea style={TitleInput} placeholder="Enter Title" />
+      </Html>
 
       {/* //* container content text area */}
-      <ContainerContent />
+      <Html>
+        <textarea style={ContentInput} placeholder="Enter content" />
+      </Html>
 
-      <ContainerFooter />
+      {/* //* container content text area */}
+      <Group
+        onMouseEnter={(e) => {
+          const container = e.target.getStage().container();
+          container.style.cursor = "pointer";
+        }}
+        onMouseLeave={(e) => {
+          const container = e.target.getStage().container();
+          container.style.cursor = "default";
+        }}
+      >
+        {/* //* drag */}
+        <Rect x={85} y={235} width={28} height={17} />
+        {/* //* top row */}
+        <Circle x={89} y={239} radius={2.5} fill="white" />
+        <Circle x={99} y={239} radius={2.5} fill="white" />
+        <Circle x={109} y={239} radius={2.5} fill="white" />
+
+        {/* //* bottom row */}
+        <Circle x={89} y={249} radius={2.5} fill="white" />
+        <Circle x={99} y={249} radius={2.5} fill="white" />
+        <Circle x={109} y={249} radius={2.5} fill="white" />
+      </Group>
     </Group>
   );
 }
